+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 mative-tweak (0.0.1) RELEASED; urgency=high
 
+  * - Structuring of the script has been carried out. Added text formatting, bolded headings and added color. Added two mative tools. Mative-fetch: analog of screenfetch. Mative-uCareSystem: Intsrument for complete system upgrade, update flatpak packages, clean the system from garbage - residual files, remove unused dependencies, upgrade kernel and remove old kernels. Added tool for installing Nouveau Nvidia - free open source driver. Added tool to update kernel from mainline branch.
+    
+ -- mative-tweak <titenko.m@gmail.com>  Sun, 19 Nov 2023 14:58:13 +0100
+
+mative-tweak (0.0.1) RELEASED; urgency=high
+
   * - Ubuntu Mainline Kernel installation has been added.
     - Installing the kernel from PPA
     - Installing kernel from developer repository
     - Added installation of open source Nvidia driver - Nouveau
     
  -- mative-tweak <titenko.m@gmail.com>  Sun, 12 Nov 2023 02:30:55 +0100
-
-